requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
beautifulsoup4==4.10.0
black==21.9b0
certifi==2021.10.8
charset-normalizer==2.0.6
click==8.0.2
dataclasses==0.8
idna==3.2
importlib-metadata==4.8.1
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.9.0
pkg-resources==0.0.0
platformdirs==2.4.0
regex==2021.10.8
requests==2.26.0
requests-oauthlib==1.3.0
soupsieve==2.2.1
tomli==1.2.1
tweepy==4.1.0
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.26.7
zipp==3.6.0
